dags/requirements.txt
apache-airflow==2.9.2
mage_ai==0.9.72
matplotlib==3.9.0
scikit_learn==1.5.0
